Separate out market details calculations
