+++ src/hooks/useUniswapMarketDetails.js
@@ -1,0 +1,281 @@
+import { get, isNil } from 'lodash';
+import { useCallback } from 'react';
+import { calculateTradeDetails } from '../handlers/uniswap';
+import {
+  convertAmountFromNativeValue,
+  convertNumberToString,
+  convertRawAmountToDecimalFormat,
+  greaterThanOrEqualTo,
+  isZero,
+  updatePrecisionToDisplay,
+} from '../helpers/utilities';
+import { logger } from '../utils';
+import useAccountSettings from './useAccountSettings';
+import useUniswapAllowances from './useUniswapAllowances';
+import useUniswapMarketPrice from './useUniswapMarketPrice';
+
+const DEFAULT_NATIVE_INPUT_AMOUNT = 50;
+
+export default function useUniswapMarketDetails() {
+  const { chainId } = useAccountSettings();
+  const { getMarketPrice } = useUniswapMarketPrice();
+
+  const { inputReserve, outputReserve } = useUniswapAllowances();
+
+  const updateTradeDetails = useCallback(
+    ({
+      inputAmount,
+      inputAsExactAmount,
+      inputCurrency,
+      outputAmount,
+      outputCurrency,
+    }) => {
+      let updatedInputAmount = inputAmount;
+      let updatedInputAsExactAmount = inputAsExactAmount;
+      const isMissingAmounts = !inputAmount && !outputAmount;
+
+      if (isMissingAmounts) {
+        const inputNativePrice = getMarketPrice(inputCurrency, outputCurrency);
+        updatedInputAmount = convertAmountFromNativeValue(
+          DEFAULT_NATIVE_INPUT_AMOUNT,
+          inputNativePrice,
+          inputCurrency.decimals
+        );
+        updatedInputAsExactAmount = true;
+      }
+
+      return calculateTradeDetails(
+        chainId,
+        updatedInputAmount,
+        inputCurrency,
+        inputReserve,
+        outputAmount,
+        outputCurrency,
+        outputReserve,
+        updatedInputAsExactAmount
+      );
+    },
+    [chainId, getMarketPrice, inputReserve, outputReserve]
+  );
+
+  const calculateInputGivenOutputChange = useCallback(
+    ({
+      inputAsExactAmount,
+      inputBalance,
+      inputCurrency,
+      inputDecimals,
+      isOutputEmpty,
+      isOutputZero,
+      setIsSufficientBalance,
+      tradeDetails,
+      updateInputAmount,
+    }) => {
+      if (isOutputEmpty || isOutputZero) {
+        updateInputAmount();
+        setIsSufficientBalance(true);
+      } else {
+        const updatedInputAmount = get(tradeDetails, 'inputAmount.amount');
+        const rawUpdatedInputAmount = convertRawAmountToDecimalFormat(
+          updatedInputAmount,
+          inputDecimals
+        );
+
+        const updatedInputAmountDisplay = updatePrecisionToDisplay(
+          rawUpdatedInputAmount,
+          get(inputCurrency, 'price.value'),
+          true
+        );
+        updateInputAmount(
+          rawUpdatedInputAmount,
+          updatedInputAmountDisplay,
+          inputAsExactAmount
+        );
+
+        const isSufficientAmountToTrade = greaterThanOrEqualTo(
+          inputBalance,
+          rawUpdatedInputAmount
+        );
+        setIsSufficientBalance(isSufficientAmountToTrade);
+      }
+    },
+    []
+  );
+
+  const calculateOutputGivenInputChange = useCallback(
+    ({
+      inputAsExactAmount,
+      inputCurrency,
+      isInputEmpty,
+      isInputZero,
+      outputCurrency,
+      outputDecimals,
+      outputFieldRef,
+      tradeDetails,
+      updateOutputAmount,
+    }) => {
+      logger.log('calculate OUTPUT given INPUT change');
+      if (
+        (isInputEmpty || isInputZero) &&
+        outputFieldRef &&
+        outputFieldRef.current &&
+        !outputFieldRef.current.isFocused()
+      ) {
+        updateOutputAmount(null, null, true);
+      } else {
+        const updatedOutputAmount = get(tradeDetails, 'outputAmount.amount');
+        const rawUpdatedOutputAmount = convertRawAmountToDecimalFormat(
+          updatedOutputAmount,
+          outputDecimals
+        );
+        if (!isZero(rawUpdatedOutputAmount)) {
+          let outputNativePrice = get(outputCurrency, 'price.value', null);
+          if (isNil(outputNativePrice)) {
+            outputNativePrice = getMarketPrice(
+              inputCurrency,
+              outputCurrency,
+              false
+            );
+          }
+          const updatedOutputAmountDisplay = updatePrecisionToDisplay(
+            rawUpdatedOutputAmount,
+            outputNativePrice
+          );
+
+          updateOutputAmount(
+            rawUpdatedOutputAmount,
+            updatedOutputAmountDisplay,
+            inputAsExactAmount
+          );
+        }
+      }
+    },
+    [getMarketPrice]
+  );
+
+  const getMarketDetails = useCallback(
+    ({
+      clearForm,
+      inputAmount,
+      inputAsExactAmount,
+      inputBalance,
+      inputCurrency,
+      inputFieldRef,
+      isMax,
+      nativeAmount,
+      nativeCurrency,
+      nativeFieldRef,
+      outputAmount,
+      outputCurrency,
+      outputFieldRef,
+      setIsSufficientBalance,
+      setSlippage,
+      updateExtraTradeDetails,
+      updateInputAmount,
+      updateOutputAmount,
+    }) => {
+      const isMissingCurrency = !inputCurrency || !outputCurrency;
+      const isMissingReserves =
+        (inputCurrency && inputCurrency.address !== 'eth' && !inputReserve) ||
+        (outputCurrency && outputCurrency.address !== 'eth' && !outputReserve);
+      if (isMissingCurrency || isMissingReserves) return;
+
+      try {
+        const tradeDetails = updateTradeDetails({
+          inputAmount,
+          inputAsExactAmount,
+          inputCurrency,
+          outputAmount,
+          outputCurrency,
+        });
+        updateExtraTradeDetails({
+          inputCurrency,
+          nativeCurrency,
+          outputCurrency,
+          tradeDetails,
+        });
+
+        const isMissingAmounts = !inputAmount && !outputAmount;
+        if (isMissingAmounts) return;
+
+        const { decimals: inputDecimals } = inputCurrency;
+        const { decimals: outputDecimals } = outputCurrency;
+
+        // update slippage
+        const slippage = convertNumberToString(
+          get(tradeDetails, 'executionRateSlippage', 0)
+        );
+        setSlippage(slippage);
+
+        const newIsSufficientBalance =
+          !inputAmount || greaterThanOrEqualTo(inputBalance, inputAmount);
+
+        setIsSufficientBalance(newIsSufficientBalance);
+
+        const isInputEmpty = !inputAmount;
+        const isNativeEmpty = !nativeAmount;
+        const isOutputEmpty = !outputAmount;
+
+        const isInputZero = Number(inputAmount) === 0;
+        const isOutputZero = Number(outputAmount) === 0;
+
+        if (
+          nativeFieldRef &&
+          nativeFieldRef.current &&
+          nativeFieldRef.current.isFocused() &&
+          isNativeEmpty &&
+          !isMax
+        ) {
+          clearForm();
+        }
+
+        // update output amount given input amount changes
+        if (inputAsExactAmount) {
+          calculateOutputGivenInputChange({
+            inputAsExactAmount,
+            inputCurrency,
+            isInputEmpty,
+            isInputZero,
+            outputCurrency,
+            outputDecimals,
+            outputFieldRef,
+            tradeDetails,
+            updateOutputAmount,
+          });
+        }
+
+        // update input amount given output amount changes
+        if (
+          !inputAsExactAmount &&
+          inputFieldRef &&
+          inputFieldRef.current &&
+          !inputFieldRef.current.isFocused()
+        ) {
+          calculateInputGivenOutputChange({
+            inputAsExactAmount,
+            inputBalance,
+            inputCurrency,
+            inputDecimals,
+            isOutputEmpty,
+            isOutputZero,
+            setIsSufficientBalance,
+            tradeDetails,
+            updateInputAmount,
+          });
+        }
+      } catch (error) {
+        logger.log('error getting market details', error);
+      }
+    },
+    [
+      calculateInputGivenOutputChange,
+      calculateOutputGivenInputChange,
+      inputReserve,
+      outputReserve,
+      updateTradeDetails,
+    ]
+  );
+
+  return {
+    getMarketDetails,
+  };
+}
